+++ src/main/kotlin/icu/windea/pls/lang/intentions/script/ReplaceArrayDefineReferenceWithEvaluationResultIntention.kt
@@ -6,19 +6,19 @@ import com.intellij.modcommand.PsiUpdateModCommandAction
 import com.intellij.psi.PsiElement
 import icu.windea.pls.PlsBundle
 import icu.windea.pls.lang.psi.ParadoxExpressionElement
-import icu.windea.pls.lang.util.evaluators.ParadoxDefineReferenceEvaluator
+import icu.windea.pls.lang.util.evaluators.ParadoxArrayDefineReferenceEvaluator
 import icu.windea.pls.lang.util.evaluators.ParadoxEvaluationService
 import icu.windea.pls.script.psi.ParadoxScriptElementFactory
 import icu.windea.pls.script.psi.ParadoxScriptValue
 
 /**
- * 将定值引用表达式所在的表达式替换为其求值结果。
+ * 将数组定值引用表达式所在的表达式替换为其求值结果。
  *
- * @see ParadoxDefineReferenceEvaluator
+ * @see ParadoxArrayDefineReferenceEvaluator
  */
 @Suppress("UnstableApiUsage")
-class ReplaceDefineReferenceWithEvaluatedValueIntention : PsiUpdateModCommandAction<ParadoxExpressionElement>(ParadoxExpressionElement::class.java) {
-    override fun getFamilyName() = PlsBundle.message("intention.replaceInlineMathWithEvaluationResult")
+class ReplaceArrayDefineReferenceWithEvaluationResultIntention : PsiUpdateModCommandAction<ParadoxExpressionElement>(ParadoxExpressionElement::class.java) {
+    override fun getFamilyName() = PlsBundle.message("intention.replaceArrayDefineReferenceWithEvaluationResult")
 
     override fun invoke(context: ActionContext, element: ParadoxExpressionElement, updater: ModPsiUpdater) {
         val result = getResult(element) ?: return
@@ -35,9 +35,9 @@ class ReplaceDefineReferenceWithEvaluatedValueIntention : PsiUpdateModCommandAct
     }
 
     private fun getResult(element: ParadoxExpressionElement): ParadoxScriptValue? {
-        if (!ParadoxEvaluationService.isEvaluableForDefineReference(element)) return null
+        if (!ParadoxEvaluationService.isEvaluableForArrayDefineReference(element)) return null
 
-        val evaluator = ParadoxDefineReferenceEvaluator(resolve = false) // NOTE 2.1.10 do not resolve scripted variables here
-        return evaluator.evaluateFromRoot(element)
+        val evaluator = ParadoxArrayDefineReferenceEvaluator(resolve = false) // NOTE 2.1.10 do not resolve scripted variables here
+        return runCatching { evaluator.evaluateFromRoot(element) }.getOrNull()
     }
 }