+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 - [x] [CK3/VIC3/EU5] 完善对定值引用表达式的支持 / Improve support for define reference expressions ([#341](https://github.com/DragonKnightOfBreeze/Paradox-Language-Support/issues/341))
 - [x] [CK3/VIC3/EU5] 完善对数组定值引用表达式的支持 / Improve support for array define reference expressions ([#341](https://github.com/DragonKnightOfBreeze/Paradox-Language-Support/issues/341))
 - [ ] [CK3/VIC3/EU5] 完善对标签集合表达式的支持 / Improve support for tags expressions ([#163](https://github.com/DragonKnightOfBreeze/Paradox-Language-Support/issues/163))
+- [x] 提供内嵌提示，以显示（数组）定值引用的评估结果 / Provide inlay hints to show evaluation results of (array) define references
+- [x] 提供意向，以将（数组）定值引用替换为评估结果 / Provide intentions to replace (array) define references with evaluation results
 - [x] 在推断游戏类型时，同时提供描述信息，并在模组设置对话框中显示 / Provide description info when inferring game type, and display it in the mod settings dialog
 - [ ] 提供意向和代码检查，以在作用域调用语句的安全形式（`owner ?= ...` 或 `owner? = ...`）与显式形式（`exists = owner owner = ...`）之间进行转换 / Provide intentions and inspections to convert scope call statements between safe form (`owner ?= ...` or `owner? = ...`) and explicit form (`exists = owner owner = ...`)
 - [x] 提供意向和代码检查，以在作用域调用语句的链式形式（`root.owner = ...`）与嵌套形式（`root = { owner = ... }`）之间进行转换 / Provide intentions and inspections to convert scope call statements between chained form (`root.owner = ...`) and nested form (`root = { owner = ... }`)